Prune files in async stream instead of sync vector
