+++ server/src/storage/compressed_data_manager.rs
@@ -151,13 +151,21 @@ impl CompressedDataManager {
     }
 
     /// Return an [`ObjectMeta`] for each compressed file in `query_data_folder` that belongs to the
-    /// table with `table_name`. If some compressed data that belongs to `table_name` is still in
-    /// memory, save it to disk first. If no files belongs to the table with `table_name` an empty
-    /// [`Vec`] is returned, while a [`DataRetrievalError`](ModelarDbError::DataRetrievalError) is
-    /// returned if the compressed files could not be listed.
+    /// table with `table_name` and contains compressed segments within the given range of time and
+    /// value. If some compressed data that belongs to `table_name` is still in memory, save it to
+    /// disk first. If no files belongs to the table with `table_name` an empty [`Vec`] is returned,
+    /// while a [`DataRetrievalError`](ModelarDbError::DataRetrievalError) is returned if:
+    /// * A table with `table_name` does not exist.
+    /// * The compressed files could not be listed.
+    /// * The end time is before the start time.
+    /// * The max value is larger than the min value.
     pub(super) async fn save_and_get_saved_compressed_files(
         &mut self,
         table_name: &str,
+        start_time: Option<Timestamp>,
+        end_time: Option<Timestamp>,
+        min_value: Option<Value>,
+        max_value: Option<Value>,
         query_data_folder: &Arc<dyn ObjectStore>,
     ) -> Result<Vec<ObjectMeta>, ModelarDbError> {
         // If there is any compressed data in memory, save it first.
@@ -175,6 +183,27 @@ impl CompressedDataManager {
                 .map_err(|error| ModelarDbError::DataRetrievalError(error.to_string()))?;
         }
 
+        // Set default values for the parts of the time and value range that is not defined.
+        let start_time = start_time.unwrap_or(0);
+        let end_time = end_time.unwrap_or(Timestamp::MAX);
+
+        if start_time > end_time {
+            return Err(ModelarDbError::DataRetrievalError(format!(
+                "Start time '{}' cannot be after end time '{}'.",
+                start_time, end_time
+            )));
+        };
+
+        let min_value = min_value.unwrap_or(Value::NEG_INFINITY);
+        let max_value = max_value.unwrap_or(Value::INFINITY);
+
+        if min_value > max_value {
+            return Err(ModelarDbError::DataRetrievalError(format!(
+                "Min value '{}' cannot be larger than max value '{}'.",
+                min_value, max_value
+            )));
+        };
+
         // List all files in query_data_folder for the table named table_name.
         let table_path = ObjectStorePath::from(format!("compressed/{}/", table_name));
         let table_files = query_data_folder
@@ -187,25 +216,27 @@ impl CompressedDataManager {
                 ))
             })?;
 
-        // Return all Apache Parquet files stored for the table named table_name.
-        let table_files = table_files
+        // Return all relevant Apache Parquet files stored for the table named table_name.
+        let table_relevant_apache_parquet_files = table_files
             .filter_map(|maybe_meta| async {
-                if let Ok(meta) = maybe_meta {
-                    if StorageEngine::is_path_an_apache_parquet_file(
+                if let Ok(object_meta) = maybe_meta {
+                    is_object_meta_relevant(
+                        object_meta,
+                        start_time,
+                        end_time,
+                        min_value,
+                        max_value,
                         query_data_folder,
-                        &meta.location,
                     )
                     .await
-                    {
-                        return Some(meta);
-                    };
-                };
-                None
+                } else {
+                    None
+                }
             })
             .collect::<Vec<ObjectMeta>>()
             .await;
 
-        Ok(table_files)
+        Ok(table_relevant_apache_parquet_files)
     }
 
     /// Save [`CompressedDataBuffers`](CompressedDataBuffer) to disk until the reserved memory limit
@@ -258,12 +289,44 @@ impl CompressedDataManager {
     }
 }
 
+/// Return the [`ObjectMeta`] in `object_meta` if it represents an Apache Parquet file and if the
+/// timestamps and values in its filename overlaps with the time range given by `start_time` and
+/// `end_time` and the value range given by `min_value` and `max_value`, otherwise [`None`] is
+/// returned. Assumes the name of Apache Parquet files represented by `object_meta` has the
+/// following format: `start_timestamp-end_timestamp-min_value-max_value.parquet`, where both
+/// `start_timestamp` and `end_timestamp` are of the same unit as `start_time` and `end_time`.
+async fn is_object_meta_relevant(
+    object_meta: ObjectMeta,
+    start_time: Timestamp,
+    end_time: Timestamp,
+    min_value: Value,
+    max_value: Value,
+    query_data_folder: &Arc<dyn ObjectStore>,
+) -> Option<ObjectMeta> {
+    let location = &object_meta.location;
+    if StorageEngine::is_path_an_apache_parquet_file(query_data_folder, location).await {
+        // unwrap() is safe as the file is created by the storage engine.
+        let last_location_part = location.parts().last().unwrap();
+
+        if is_compressed_file_within_time_and_value_range(
+            last_location_part.as_ref(),
+            start_time,
+            end_time,
+            min_value,
+            max_value,
+        ) {
+            return Some(object_meta);
+        }
+    }
+    None
+}
+
 /// Return [`true`] if the timestamps and values in `file name` overlaps with the time range given
 /// by `start_time` and `end_time` and the value range given by `min_value` and `max_value`,
 /// otherwise [`false`]. Assumes `file_name` has the following format:
 /// `start_timestamp-end_timestamp-min_value-max_value.parquet`, where both `start_timestamp` and
 /// `end_timestamp` are of the same unit as `start_time` and `end_time`.
-pub(super) fn is_compressed_file_within_time_and_value_range(
+fn is_compressed_file_within_time_and_value_range(
     file_name: &str,
     start_time: Timestamp,
     end_time: Timestamp,
@@ -298,15 +361,18 @@ mod tests {
     use super::*;
     use std::path::Path;
 
+    use datafusion::arrow::compute;
     use object_store::local::LocalFileSystem;
     use tempfile::{tempdir, TempDir};
 
+    use crate::get_array;
     use crate::metadata::test_util as metadata_test_util;
-    use crate::storage;
-    use crate::storage::test_util;
+    use crate::storage::{self, test_util};
+    use crate::types::{TimestampArray, ValueArray};
 
     const TABLE_NAME: &str = "table";
 
+    // Tests for insert_record_batch().
     #[tokio::test]
     async fn test_insert_record_batch() {
         let record_batch = test_util::get_compressed_segments_record_batch();
@@ -337,6 +403,7 @@ mod tests {
         assert_eq!(table_folder_files.len(), 1);
     }
 
+    // Tests for insert_compressed_segments().
     #[test]
     fn test_can_insert_compressed_segment_into_new_compressed_data_buffer() {
         let segments = test_util::get_compressed_segments_record_batch();
@@ -457,7 +524,350 @@ mod tests {
         )
     }
 
-    // Tests for get_saved_compressed_files().
+    // Tests for save_and_get_saved_compressed_files().
+    #[tokio::test]
+    async fn test_can_get_compressed_file_for_table() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+
+        // Insert compressed segments into the same table.
+        let segment = test_util::get_compressed_segments_record_batch();
+        data_manager
+            .insert_compressed_segments(TABLE_NAME, segment.clone())
+            .unwrap();
+        data_manager
+            .insert_compressed_segments(TABLE_NAME, segment)
+            .unwrap();
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager
+            .save_and_get_saved_compressed_files(TABLE_NAME, None, None, None, None, &object_store)
+            .await;
+
+        assert!(result.is_ok());
+        assert_eq!(result.unwrap().len(), 1);
+    }
+
+    #[tokio::test]
+    async fn test_get_no_compressed_files_for_non_existent_table() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            None,
+            None,
+            None,
+            None,
+            &object_store,
+        );
+
+        assert!(result.await.unwrap().is_empty());
+    }
+
+    #[tokio::test]
+    async fn test_can_get_compressed_file_with_start_time() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+        let (segment_1, _segment_2) = insert_separated_segments(&mut data_manager, 0, 0.0);
+
+        // If we have a start time after the first segments ends, only the file containing the
+        // second segment should be retrieved.
+        let end_times = get_array!(segment_1, 3, TimestampArray);
+        let start_time = Some(end_times.value(end_times.len() - 1) + 100);
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            start_time,
+            None,
+            None,
+            None,
+            &object_store,
+        );
+        let files = result.await.unwrap();
+        assert_eq!(files.len(), 1);
+
+        // The path to the returned compressed file should contain the table name, the start time,
+        // end time, min value, and max value of the second segment.
+        let file_path = files.get(0).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
+        );
+    }
+
+    #[tokio::test]
+    async fn test_can_get_compressed_file_with_min_value() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+        let (segment_1, _segment_2) = insert_separated_segments(&mut data_manager, 0, 0.0);
+
+        // If we have a min value higher then the max value in the first segment, only the file
+        // containing the second segment should be retrieved.
+        let max_values = get_array!(segment_1, 6, ValueArray);
+        let min_value = Some(compute::max(max_values).unwrap() + 1.0);
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            None,
+            None,
+            min_value,
+            None,
+            &object_store,
+        );
+        let files = result.await.unwrap();
+        assert_eq!(files.len(), 1);
+
+        // The path to the returned compressed file should contain the table name, the start time,
+        // end time, min value, and max value of the second segment.
+        let file_path = files.get(0).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
+        );
+    }
+
+    #[tokio::test]
+    async fn test_can_get_compressed_file_with_end_time() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+        let (_segment_1, segment_2) = insert_separated_segments(&mut data_manager, 0, 0.0);
+
+        // If we have an end time before the second segment starts, only the file containing the
+        // first segment should be retrieved.
+        let start_times = get_array!(segment_2, 2, TimestampArray);
+        let end_time = Some(start_times.value(1) - 100);
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            None,
+            end_time,
+            None,
+            None,
+            &object_store,
+        );
+        let files = result.await.unwrap();
+        assert_eq!(files.len(), 1);
+
+        // The path to the returned compressed file should contain the table name, the start time,
+        // end time, min value, and max value of the first segment.
+        let file_path = files.get(0).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/1000_1005_5.2_34.2.parquet", TABLE_NAME)
+        );
+    }
+
+    #[tokio::test]
+    async fn test_can_get_compressed_file_with_max_value() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+        let (_segment_1, segment_2) = insert_separated_segments(&mut data_manager, 0, 0.0);
+
+        // If we have a max value lower then the min value in the second segment, only the file
+        // containing the first segment should be retrieved.
+        let min_values = get_array!(segment_2, 5, ValueArray);
+        let max_value = Some(compute::min(min_values).unwrap() - 1.0);
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            None,
+            None,
+            None,
+            max_value,
+            &object_store,
+        );
+        let files = result.await.unwrap();
+        assert_eq!(files.len(), 1);
+
+        // The path to the returned compressed file should contain the table name, the start time,
+        // end time, min value, and max value of the first segment.
+        let file_path = files.get(0).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/1000_1005_5.2_34.2.parquet", TABLE_NAME)
+        );
+    }
+
+    #[tokio::test]
+    async fn test_can_get_compressed_files_with_start_time_and_end_time() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+
+        // Insert 4 segments with a ~1 second time difference between segment 1 and 2 and segment 3 and 4.
+        let (segment_1, _segment_2) = insert_separated_segments(&mut data_manager, 0, 0.0);
+        let (_segment_3, segment_4) = insert_separated_segments(&mut data_manager, 1000, 0.0);
+
+        // If we have a start time after the first segment and an end time before the fourth
+        // segment, only the files containing the second and third segment should be retrieved.
+        let end_times = get_array!(segment_1, 3, TimestampArray);
+        let start_times = get_array!(segment_4, 2, TimestampArray);
+
+        let start_time = Some(end_times.value(end_times.len() - 1) + 100);
+        let end_time = Some(start_times.value(1) - 100);
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            start_time,
+            end_time,
+            None,
+            None,
+            &object_store,
+        );
+        let mut files = result.await.unwrap();
+        assert_eq!(files.len(), 2);
+
+        // Sort files as save_and_get_saved_compressed_files() does not guarantee an ordering.a
+        files.sort_unstable_by(|a, b| a.location.cmp(&b.location));
+
+        // The files containing the second and third segment should be returned and their path
+        // should contain the table name, the start time, end time, min value, and max value of the
+        // segment in each file.
+        let file_path = files.get(0).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
+        );
+        let file_path = files.get(1).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/2000_2005_5.2_34.2.parquet", TABLE_NAME)
+        );
+    }
+
+    #[tokio::test]
+    async fn test_can_get_compressed_files_with_min_value_and_max_value() {
+        let (temp_dir, mut data_manager) = create_compressed_data_manager();
+
+        // Insert 4 segments with a ~1 second time difference between segment 1 and 2 and segment 3 and 4.
+        let (segment_1, _segment_2) = insert_separated_segments(&mut data_manager, 0, 0.0);
+        let (_segment_3, segment_4) = insert_separated_segments(&mut data_manager, 1000, 100.0);
+
+        // If we have a min value higher the first segment and a max value lower than the fourth
+        // segment, only the files containing the second and third segment should be retrieved.
+        let max_values = get_array!(segment_1, 6, ValueArray);
+        let min_values = get_array!(segment_4, 5, ValueArray);
+
+        let min_value = Some(compute::max(max_values).unwrap() + 1.0);
+        let max_value = Some(compute::min(min_values).unwrap() - 1.0);
+
+        let object_store: Arc<dyn ObjectStore> =
+            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            None,
+            None,
+            min_value,
+            max_value,
+            &object_store,
+        );
+        let mut files = result.await.unwrap();
+        assert_eq!(files.len(), 2);
+
+        // Sort files as save_and_get_saved_compressed_files() does not guarantee an ordering.
+        files.sort_unstable_by(|a, b| a.location.cmp(&b.location));
+
+        // The files containing the second and third segment should be returned and their path
+        // should contain the table name, the start time, end time, min value, and max value of the
+        // segment in each file.
+        let file_path = files.get(0).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/2000_2005_105.2_134.2.parquet", TABLE_NAME)
+        );
+        let file_path = files.get(1).unwrap().location.to_string();
+        assert_eq!(
+            file_path,
+            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
+        );
+    }
+
+    /// Create and insert two compressed segments with a 1 second time difference offset by
+    /// `start_time` and values with difference of 10 offset by `value_offset`.
+    fn insert_separated_segments(
+        data_manager: &mut CompressedDataManager,
+        start_time: i64,
+        value_offset: f32,
+    ) -> (RecordBatch, RecordBatch) {
+        let segment_1 = test_util::get_compressed_segments_record_batch_with_time(
+            1000 + start_time,
+            value_offset,
+        );
+        data_manager
+            .insert_compressed_segments(TABLE_NAME, segment_1.clone())
+            .unwrap();
+        data_manager.flush().unwrap();
+
+        let segment_2 = test_util::get_compressed_segments_record_batch_with_time(
+            2000 + start_time,
+            10.0 + value_offset,
+        );
+        data_manager
+            .insert_compressed_segments(TABLE_NAME, segment_2.clone())
+            .unwrap();
+        data_manager.flush().unwrap();
+
+        (segment_1, segment_2)
+    }
+
+    #[tokio::test]
+    async fn test_cannot_get_compressed_files_where_end_time_is_before_start_time() {
+        let (_temp_dir, mut data_manager) = create_compressed_data_manager();
+
+        let segment = test_util::get_compressed_segments_record_batch();
+        data_manager
+            .insert_compressed_segments(TABLE_NAME, segment)
+            .unwrap();
+
+        let object_store: Arc<dyn ObjectStore> = Arc::new(LocalFileSystem::new());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            Some(10),
+            Some(1),
+            None,
+            None,
+            &object_store,
+        );
+
+        assert!(result
+            .await
+            .unwrap_err()
+            .to_string()
+            .contains("cannot be after end time"));
+    }
+
+    #[tokio::test]
+    async fn test_cannot_get_compressed_files_where_max_value_is_smaller_than_min_value() {
+        let (_temp_dir, mut data_manager) = create_compressed_data_manager();
+
+        let segment = test_util::get_compressed_segments_record_batch();
+        data_manager
+            .insert_compressed_segments(TABLE_NAME, segment)
+            .unwrap();
+
+        let object_store: Arc<dyn ObjectStore> = Arc::new(LocalFileSystem::new());
+        let result = data_manager.save_and_get_saved_compressed_files(
+            TABLE_NAME,
+            None,
+            None,
+            Some(10.0),
+            Some(1.0),
+            &object_store,
+        );
+
+        assert!(result
+            .await
+            .unwrap_err()
+            .to_string()
+            .contains("cannot be larger than max value"));
+    }
+
     #[tokio::test]
     async fn test_can_get_saved_compressed_files() {
         let segments = test_util::get_compressed_segments_record_batch();
@@ -471,7 +881,7 @@ mod tests {
         let object_store: Arc<dyn ObjectStore> =
             Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
         let result = data_manager
-            .save_and_get_saved_compressed_files(TABLE_NAME, &object_store)
+            .save_and_get_saved_compressed_files(TABLE_NAME, None, None, None, None, &object_store)
             .await;
         assert!(result.is_ok());
 
@@ -507,7 +917,7 @@ mod tests {
         let object_store: Arc<dyn ObjectStore> =
             Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
         let result = data_manager
-            .save_and_get_saved_compressed_files(TABLE_NAME, &object_store)
+            .save_and_get_saved_compressed_files(TABLE_NAME, None, None, None, None, &object_store)
             .await;
 
         assert!(result.is_ok());
@@ -527,7 +937,7 @@ mod tests {
         let object_store: Arc<dyn ObjectStore> =
             Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
         let result = data_manager
-            .save_and_get_saved_compressed_files("NO_TABLE", &object_store)
+            .save_and_get_saved_compressed_files("NO_TABLE", None, None, None, None, &object_store)
             .await;
         assert!(result.unwrap().is_empty());
     }
