+++ server/src/storage/mod.rs
@@ -235,49 +235,16 @@ impl StorageEngine {
         max_value: Option<Value>,
         query_data_folder: &Arc<dyn ObjectStore>,
     ) -> Result<Vec<ObjectMeta>, ModelarDbError> {
-        let start_time = start_time.unwrap_or(0);
-        let end_time = end_time.unwrap_or(Timestamp::MAX);
-
-        if start_time > end_time {
-            return Err(ModelarDbError::DataRetrievalError(format!(
-                "Start time '{}' cannot be after end time '{}'.",
-                start_time, end_time
-            )));
-        };
-
-        let min_value = min_value.unwrap_or(Value::NEG_INFINITY);
-        let max_value = max_value.unwrap_or(Value::INFINITY);
-
-        if min_value > max_value {
-            return Err(ModelarDbError::DataRetrievalError(format!(
-                "Min value '{}' cannot be larger than max value '{}'.",
-                min_value, max_value
-            )));
-        };
-
-        // List the files storing the compressed data for the table with table_name.
-        let compressed_files = self
-            .compressed_data_manager
-            .save_and_get_saved_compressed_files(table_name, query_data_folder)
-            .await?;
-
-        // TODO: prune by time and value in save_and_get_saved_compressed_files to perform the action async?
-        // Prune files based on the time and value range each file contains compressed segments for.
-        let pruned_compressed_files = compressed_files
-            .into_iter()
-            .filter(|object_meta| {
-                let last_file_part = object_meta.location.parts().last().unwrap();
-                compressed_data_manager::is_compressed_file_within_time_and_value_range(
-                    last_file_part.as_ref(),
-                    start_time,
-                    end_time,
-                    min_value,
-                    max_value,
-                )
-            })
-            .collect();
-
-        Ok(pruned_compressed_files)
+        self.compressed_data_manager
+            .save_and_get_saved_compressed_files(
+                table_name,
+                start_time,
+                end_time,
+                min_value,
+                max_value,
+                query_data_folder,
+            )
+            .await
     }
 
     /// Write `batch` to an Apache Parquet file at the location given by `file_path`. `file_path`
@@ -386,374 +353,9 @@ mod tests {
     use std::path::PathBuf;
     use std::sync::Arc;
 
-    use datafusion::arrow::compute;
     use datafusion::arrow::datatypes::Schema;
     use object_store::local::LocalFileSystem;
     use tempfile::{tempdir, TempDir};
-
-    use crate::get_array;
-    use crate::metadata::test_util as metadata_test_util;
-    use crate::types::TimestampArray;
-
-    const TABLE_NAME: &str = "table";
-
-    // Tests for get_compressed_files().
-    #[tokio::test]
-    async fn test_can_get_compressed_file_for_table() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-
-        // Insert compressed segments into the same table.
-        let segment = test_util::get_compressed_segments_record_batch();
-        storage_engine
-            .compressed_data_manager
-            .insert_compressed_segments(TABLE_NAME, segment.clone())
-            .unwrap();
-        storage_engine
-            .compressed_data_manager
-            .insert_compressed_segments(TABLE_NAME, segment)
-            .unwrap();
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine
-            .get_compressed_files(TABLE_NAME, None, None, None, None, &object_store)
-            .await;
-
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_get_no_compressed_files_for_non_existent_table() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result =
-            storage_engine.get_compressed_files(TABLE_NAME, None, None, None, None, &object_store);
-
-        assert!(result.await.unwrap().is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_can_get_compressed_file_with_start_time() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-        let (segment_1, _segment_2) = insert_separated_segments(&mut storage_engine, 0, 0.0);
-
-        // If we have a start time after the first segments ends, only the file containing the
-        // second segment should be retrieved.
-        let end_times = get_array!(segment_1, 3, TimestampArray);
-        let start_time = Some(end_times.value(end_times.len() - 1) + 100);
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            start_time,
-            None,
-            None,
-            None,
-            &object_store,
-        );
-        let files = result.await.unwrap();
-        assert_eq!(files.len(), 1);
-
-        // The path to the returned compressed file should contain the table name, the start time,
-        // end time, min value, and max value of the second segment.
-        let file_path = files.get(0).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_can_get_compressed_file_with_min_value() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-        let (segment_1, _segment_2) = insert_separated_segments(&mut storage_engine, 0, 0.0);
-
-        // If we have a min value higher then the max value in the first segment, only the file
-        // containing the second segment should be retrieved.
-        let max_values = get_array!(segment_1, 6, ValueArray);
-        let min_value = Some(compute::max(max_values).unwrap() + 1.0);
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            None,
-            None,
-            min_value,
-            None,
-            &object_store,
-        );
-        let files = result.await.unwrap();
-        assert_eq!(files.len(), 1);
-
-        // The path to the returned compressed file should contain the table name, the start time,
-        // end time, min value, and max value of the second segment.
-        let file_path = files.get(0).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_can_get_compressed_file_with_end_time() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-        let (_segment_1, segment_2) = insert_separated_segments(&mut storage_engine, 0, 0.0);
-
-        // If we have an end time before the second segment starts, only the file containing the
-        // first segment should be retrieved.
-        let start_times = get_array!(segment_2, 2, TimestampArray);
-        let end_time = Some(start_times.value(1) - 100);
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            None,
-            end_time,
-            None,
-            None,
-            &object_store,
-        );
-        let files = result.await.unwrap();
-        assert_eq!(files.len(), 1);
-
-        // The path to the returned compressed file should contain the table name, the start time,
-        // end time, min value, and max value of the first segment.
-        let file_path = files.get(0).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/1000_1005_5.2_34.2.parquet", TABLE_NAME)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_can_get_compressed_file_with_max_value() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-        let (_segment_1, segment_2) = insert_separated_segments(&mut storage_engine, 0, 0.0);
-
-        // If we have a max value lower then the min value in the second segment, only the file
-        // containing the first segment should be retrieved.
-        let min_values = get_array!(segment_2, 5, ValueArray);
-        let max_value = Some(compute::min(min_values).unwrap() - 1.0);
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            None,
-            None,
-            None,
-            max_value,
-            &object_store,
-        );
-        let files = result.await.unwrap();
-        assert_eq!(files.len(), 1);
-
-        // The path to the returned compressed file should contain the table name, the start time,
-        // end time, min value, and max value of the first segment.
-        let file_path = files.get(0).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/1000_1005_5.2_34.2.parquet", TABLE_NAME)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_can_get_compressed_files_with_start_time_and_end_time() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-
-        // Insert 4 segments with a ~1 second time difference between segment 1 and 2 and segment 3 and 4.
-        let (segment_1, _segment_2) = insert_separated_segments(&mut storage_engine, 0, 0.0);
-        let (_segment_3, segment_4) = insert_separated_segments(&mut storage_engine, 1000, 0.0);
-
-        // If we have a start time after the first segment and an end time before the fourth
-        // segment, only the files containing the second and third segment should be retrieved.
-        let end_times = get_array!(segment_1, 3, TimestampArray);
-        let start_times = get_array!(segment_4, 2, TimestampArray);
-
-        let start_time = Some(end_times.value(end_times.len() - 1) + 100);
-        let end_time = Some(start_times.value(1) - 100);
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            start_time,
-            end_time,
-            None,
-            None,
-            &object_store,
-        );
-        let mut files = result.await.unwrap();
-        assert_eq!(files.len(), 2);
-
-        // Sort files as get_compressed_files() does not guarantee an ordering.
-        files.sort_unstable_by(|a, b| a.location.cmp(&b.location));
-
-        // The files containing the second and third segment should be returned and their path
-        // should contain the table name, the start time, end time, min value, and max value of the
-        // segment in each file.
-        let file_path = files.get(0).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
-        );
-        let file_path = files.get(1).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/2000_2005_5.2_34.2.parquet", TABLE_NAME)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_can_get_compressed_files_with_min_value_and_max_value() {
-        let (temp_dir, mut storage_engine) = create_storage_engine().await;
-
-        // Insert 4 segments with a ~1 second time difference between segment 1 and 2 and segment 3 and 4.
-        let (segment_1, _segment_2) = insert_separated_segments(&mut storage_engine, 0, 0.0);
-        let (_segment_3, segment_4) = insert_separated_segments(&mut storage_engine, 1000, 100.0);
-
-        // If we have a min value higher the first segment and a max value lower than the fourth
-        // segment, only the files containing the second and third segment should be retrieved.
-        let max_values = get_array!(segment_1, 6, ValueArray);
-        let min_values = get_array!(segment_4, 5, ValueArray);
-
-        let min_value = Some(compute::max(max_values).unwrap() + 1.0);
-        let max_value = Some(compute::min(min_values).unwrap() - 1.0);
-
-        let object_store: Arc<dyn ObjectStore> =
-            Arc::new(LocalFileSystem::new_with_prefix(temp_dir.path()).unwrap());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            None,
-            None,
-            min_value,
-            max_value,
-            &object_store,
-        );
-        let mut files = result.await.unwrap();
-        assert_eq!(files.len(), 2);
-
-        // Sort files as get_compressed_files() does not guarantee an ordering.
-        files.sort_unstable_by(|a, b| a.location.cmp(&b.location));
-
-        // The files containing the second and third segment should be returned and their path
-        // should contain the table name, the start time, end time, min value, and max value of the
-        // segment in each file.
-        let file_path = files.get(0).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/2000_2005_105.2_134.2.parquet", TABLE_NAME)
-        );
-        let file_path = files.get(1).unwrap().location.to_string();
-        assert_eq!(
-            file_path,
-            format!("compressed/{}/2000_2005_15.2_44.2.parquet", TABLE_NAME)
-        );
-    }
-
-    /// Create and insert two compressed segments with a 1 second time difference offset by
-    /// `start_time` and values with difference of 10 offset by `value_offset`.
-    fn insert_separated_segments(
-        storage_engine: &mut StorageEngine,
-        start_time: i64,
-        value_offset: f32,
-    ) -> (RecordBatch, RecordBatch) {
-        let segment_1 = test_util::get_compressed_segments_record_batch_with_time(
-            1000 + start_time,
-            value_offset,
-        );
-        storage_engine
-            .compressed_data_manager
-            .insert_compressed_segments(TABLE_NAME, segment_1.clone())
-            .unwrap();
-        storage_engine.compressed_data_manager.flush().unwrap();
-
-        let segment_2 = test_util::get_compressed_segments_record_batch_with_time(
-            2000 + start_time,
-            10.0 + value_offset,
-        );
-        storage_engine
-            .compressed_data_manager
-            .insert_compressed_segments(TABLE_NAME, segment_2.clone())
-            .unwrap();
-        storage_engine.compressed_data_manager.flush().unwrap();
-
-        (segment_1, segment_2)
-    }
-
-    #[tokio::test]
-    async fn test_cannot_get_compressed_files_where_end_time_is_before_start_time() {
-        let (_temp_dir, mut storage_engine) = create_storage_engine().await;
-
-        let segment = test_util::get_compressed_segments_record_batch();
-        storage_engine
-            .compressed_data_manager
-            .insert_compressed_segments(TABLE_NAME, segment)
-            .unwrap();
-
-        let object_store: Arc<dyn ObjectStore> = Arc::new(LocalFileSystem::new());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            Some(10),
-            Some(1),
-            None,
-            None,
-            &object_store,
-        );
-
-        assert!(result
-            .await
-            .unwrap_err()
-            .to_string()
-            .contains("cannot be after end time"));
-    }
-
-    #[tokio::test]
-    async fn test_cannot_get_compressed_files_where_max_value_is_smaller_than_min_value() {
-        let (_temp_dir, mut storage_engine) = create_storage_engine().await;
-
-        let segment = test_util::get_compressed_segments_record_batch();
-        storage_engine
-            .compressed_data_manager
-            .insert_compressed_segments(TABLE_NAME, segment)
-            .unwrap();
-
-        let object_store: Arc<dyn ObjectStore> = Arc::new(LocalFileSystem::new());
-        let result = storage_engine.get_compressed_files(
-            TABLE_NAME,
-            None,
-            None,
-            Some(10.0),
-            Some(1.0),
-            &object_store,
-        );
-
-        assert!(result
-            .await
-            .unwrap_err()
-            .to_string()
-            .contains("cannot be larger than max value"));
-    }
-
-    /// Create a [`StorageEngine`] with a folder that is deleted once the test is finished.
-    async fn create_storage_engine() -> (TempDir, StorageEngine) {
-        let temp_dir = tempdir().unwrap();
-        let local_data_folder = temp_dir.path().to_path_buf();
-        let metadata_manager =
-            metadata_test_util::get_test_metadata_manager(local_data_folder.as_path());
-        let storage_engine =
-            StorageEngine::try_new(local_data_folder, None, metadata_manager, false)
-                .await
-                .unwrap();
-
-        (temp_dir, storage_engine)
-    }
 
     // Tests for writing and reading Apache Parquet files.
     #[test]
